Add application step and use it
